rename files and delete subsciptions
