+++ frontend/src/app/models/user/update-user-notifications-dto.ts
